GEODE-2817: Add new function security callback

+++ geode-core/src/test/java/org/apache/geode/internal/cache/OplogFlushTest.java
@@ -85,9 +85,9 @@ public class OplogFlushTest extends DiskRegionTestingBase {
     }
   }
 
-  class FakeChannelWriteArrayBB implements Answer<Integer> {
+  class FakeChannelWriteArrayBB implements Answer<Long> {
     @Override
-    public Integer answer(InvocationOnMock invocation) throws Throwable {
+    public Long answer(InvocationOnMock invocation) throws Throwable {
       System.out.println("### in FakeChannelWriteArrayBB.answer :");
       return fakeWriteArrayBB(bbArray);
     }
@@ -96,7 +96,7 @@ public class OplogFlushTest extends DiskRegionTestingBase {
   /**
    * This method tries to write half of the byte buffer to the channel.
    */
-  private int fakeWriteArrayBB(ByteBuffer[] bbArray) throws IOException {
+  private long fakeWriteArrayBB(ByteBuffer[] bbArray) throws IOException {
     nFakeChannelWrites++;
     for (ByteBuffer b : bbArray) {
       int numFakeWrite = b.limit() / 2;
